feat(security): comprehensive P5-Sec audit pass — enable all WPCS security sniff families in CI
Re-enabled and audited per call site (47 files changed, +1203/-760):

  - WordPress.DB.PreparedSQL.* (foundation): 1 false-positive at
    woocommerce-functions.php line 463 properly suppressed with documented
    audit reason (inputs pre-escaped by WP_Meta_Query/WP_Tax_Query/esc_sql).
    Replaces legacy '// WPCS: unprepared SQL ok.' annotation that modern
    PHPCS silently ignored.
  - WordPress.Security.EscapeOutput.* (NotEscaped, UnsafePrintingFunction,
    ExceptionNotEscaped): 67 violations audited. Real esc_html/esc_attr/
    esc_url/wp_kses_post wraps where appropriate; phpcs:ignore with audit
    reason for trusted WP-core/WC-core/VC-core return values (apply_filters
    on woocommerce_loop_add_to_cart_link, lafka_post_nav, VC ob_get_clean,
    Tribe Events helpers).
  - WordPress.Security.NonceVerification.* (Missing + Recommended): 2 + 22
    violations audited. Real wp_verify_nonce check added to
    lafka_update_mega_menu_item() (nav-menu edit CSRF surface).
    phpcs:disable blocks with documented audit reasons for WC shop/cart
    filter contexts, price-filter widget URL preservation, and admin-display
    \$_GET reads.
  - WordPress.PHP.StrictInArray.MissingTrueStrict: 16 violations fixed.
    in_array() updated to strict comparison.

Real security improvements (not just suppressions):
  - CSRF fix: explicit wp_verify_nonce in lafka_update_mega_menu_item
    (hooked to wp_update_nav_menu_item; nav-menu CSRF).
  - XSS fix: wc_price() output wrapped with wp_kses_post in
    woocommerce-functions.php line 1254 (was unescaped).
  - XSS fix: \$lafka_account_holder_classes / \$lafka_single_product_gallery_classes
    now esc_attr-wrapped in attribute context.
  - XSS fix: wc_get_account_menu_item_classes output now esc_attr-wrapped.
  - XSS fix: urldecode(do_shortcode(...)) output now wp_kses_post-wrapped.
  - XSS fix: 4 \$_GET shop archive query params now properly esc_attr-wrapped.
  - 7 legacy '// WPCS:' annotations modernized to '// phpcs:ignore Sniff.Name
    -- reason' syntax with documented audit reasons.
  - Dead WC <3.6 version_compare check removed from price-filter widget.

PHPCBF auto-fix: 391 latent formatting violations cleared (function-call
signature spacing, embedded-PHP tag placement). These accumulated in
Session 8's new files (editorial partials, PDP partials, cart, single-product).

Dead exclusions removed from .phpcs.xml.dist:
  - WordPress.Security.ValidatedSanitizedInput.* (sniff family deprecated in WPCS 3.x).

Tests: 219 / 405 assertions — all green.
PHPCS: clean against the full WordPress-Extra ruleset.

+++ partials/blog-post-meta-bottom.php
@@ -52,7 +52,7 @@ if ( ( lafka_get_option( 'show_author_avatar' ) ) || ( is_singular() && ! lafka_
 			<?php the_tags( '<i class="fa fa-tags"></i> ' ); ?>
 			<span class="count_comments"><i class="fa fa-comments"></i> <a
 						href="<?php echo esc_url( get_comments_link() ); ?>"
-						title="<?php esc_attr_e( 'View comments', 'lafka' ); ?>"><?php echo get_comments_number(); ?></a></span>
+						title="<?php esc_attr_e( 'View comments', 'lafka' ); ?>"><?php echo (int) get_comments_number(); ?></a></span>
 		<?php endif; ?>
 	<?php endif; ?>
