+++ partials/blog-post-meta-top.php
@@ -25,6 +25,6 @@
 		<?php the_tags( '<i class="fa fa-tags"></i> ' ); ?>
 		<span class="count_comments"><i class="fa fa-comments"></i> <a
 					href="<?php echo esc_url( get_comments_link() ); ?>"
-					title="<?php esc_attr_e( 'View comments', 'lafka' ); ?>"><?php echo get_comments_number(); ?></a></span>
+					title="<?php esc_attr_e( 'View comments', 'lafka' ); ?>"><?php echo (int) get_comments_number(); ?></a></span>
 	<?php endif; ?>
 </div>
